requirements.txt
fastapi==0.68.0
python-jose[cryptography]==3.3.0
sqlalchemy==1.4.23
psycopg2==2.9.1
pydantic==1.8.2
dotenv==0.0.5